split heatmap geom types into separated files
